Consolidate NH tests
